Moved some class definitions into separate files.
